% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sendUDP}
\alias{sendUDP}
\title{Send a message in a UDP packet}
\usage{
sendUDP(serverIP, port, message)
}
\arguments{
\item{serverIP}{Destination IP address}

\item{port}{Destination port number}

\item{message}{Message string}
}
\value{
success_status
}
\description{
Send a message in a UDP packet
}
