% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sendUDP}
\alias{sendUDP}
\title{Send a message in a UDP packet from R. Returns status, 0 = no error}
\usage{
sendUDP(serverIP, port, message)
}
\description{
Send a message in a UDP packet from R. Returns status, 0 = no error
}
